finish p38

+++ src/main/java/com/cwjee/community/controller/PublishController.java
@@ -47,7 +47,7 @@ public class PublishController {
             @RequestParam(name = "title",required = false) String title,
             @RequestParam(name = "description", required = false) String description,
             @RequestParam(name = "tag",required = false)String tag,
-            @RequestParam(name = "id",required = false)Integer id,
+            @RequestParam(name = "id",required = false)long id,
             HttpServletRequest request,
             Model model){
